Refactor solarized-definitions.el for customization
* add macro solarized-with-color-variables
* pull solarized-apply-definitions out of create-solarized-theme

This makes it easier to customize and test proposed changes to
solarized-color-definitions. solarized-with-color-variables allows
definition of faces within the same context that is used for
solarized-color-definitions. Hence, the changes you test yourself can
be directly copied into the official code-base for a PR. They are also
easy to maintain in your own Emacs init scripts before merge or if
tweaking the official settings. The approach is copied from @bbatsov's
themes.

For example, setting up additions/customizations goes like this:

(defvar my/solarized-extra-definitions
  (solarized-with-color-variables
    `(;; cperl
      (cperl-array-face (,@fg-blue))
      (cperl-hash-face (,@fg-blue))
      (cperl-nonoverridable-face (,@fg-magenta)))))
(solarized-apply-definitions my/solarized-extra-definitions 'solarized)